requirements.txt
geopandas==1.1.1
keras-hub==0.21.1
keras-nlp==0.21.1
keras==3.10.0
matplotlib-inline==0.1.7
matplotlib-venn==1.1.2
matplotlib==3.10.0
numpy==2.0.2
pandas-datareader==0.10.0
pandas-gbq==0.29.2
pandas-stubs==2.2.2.240909
pandas==2.2.2
pillow==11.3.0
scikit-image==0.25.2
scikit-learn==1.6.1
sklearn-pandas==2.2.0
tensorflow-datasets==4.9.9
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.2
tensorflow-probability==0.25.0
tensorflow-text==2.19.0
tensorflow==2.19.0
tensorflow_decision_forests==1.12.0
tf_keras==2.19.0
tqdm==4.67.1
streamlit==1.36.0
